Add files via upload
In this code, we start by generating a synthetic dataset with two clusters of data points. Then, we perform k-means clustering using the kmeans function, specifying the number of clusters (k) and some options for the algorithm.

Finally, we plot the original data points, colored according to their assigned cluster, and mark the cluster centers with black crosses.